# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Flipeet",
      "description": "Flipeet Books E-Commerce Store Apis Documentation",
      "version": "1.0.0"
    },
    "schemes": [
      "http"
    ],
    "basePath": "/",
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/activities/reviewSeller": {
        "post": {
          "summary": "Review on Saller.",
          "description": "A User can give review to Seller.",
          "parameters": [
            {
            "name": "userid",
            "description": "User Id who send review.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "review",
            "description": "Review of User to Seller.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "bookid",
            "description": "Book Id.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "type",
            "enum": ["Book Review", "Seller Review"],
            "description": "this query tell that review on book or seller.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "sellerid",
            "description": "Book seller id",
            "type": "String",
            "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Review Add on Seller."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/plateformComments": {
        "post": {
          "summary": "Comment about flipeet.",
          "description": "A User can give any comment about flipeet.",
          "parameters": [
            {
            "name": "userid",
            "description": "User Id who send comment about flipeet.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "comment",
            "description": "Comment about flipeet.",
            "type": "String",
            "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Commented on flipeet."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/reportSeller": {
        "post": {
          "summary": "Report about seller.",
          "description": "A User can give any report of seller.",
          "parameters": [
            
            {
            "name": "reporterID",
            "description": "User Id who send report about seller.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "description",
            "description": "Description of report about Seller.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "reporteeID",
            "description": "Id of seller.",
            "type": "String",
            "in": "query"
            }
            
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Report about Seller."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/purchaseRequest": {
        "post": {
          "summary": " Book Purchase request Sent.",
          "description": "A User can sent purchase resquest of a book to seller.",
          "parameters": [
            {
            "name": "userid",
            "description": "User Id who send purchase request to seller.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "sellerid",
            "description": "Seller Id who book.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "bookid",
            "description": "Book Id whose buy a buyer.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "email",
            "description": "Email Id of buyer.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "description",
            "description": "Description about purchase request.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "status",
            "enum": ["Pending", "Accepted", "Rejected"],
            "description": "This is status of request that is used to sent request.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "chargeId",
            "description": "charge Id of Book.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "paymentMethod",
            "description": "payment method of book.",
            "type": "String",
            "in": "query"
            }
            
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Purchase request Send Successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/getPurchaseRequestBuyer": {
        "get": {
          "summary": " Sent Purchase Requests show to Seller.",
          "description": "Recieved Buyer purchase requests that is send from the User.",
          "parameters": [
            {
            "name": "sellerid",
            "description": "That is collect from loggedin user.",
            "type": "String",
            "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": " All Purchase Request that is send from buyer."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/getnewPurchaseRequestBuyer": {
        "get": {
          "summary": "Sent Purchase Request show to buyer",
          "description": "Get Buyer purchase requests that is send from the User.",
          "parameters": [
            {
            "name": "userid",
            "description": "That is collect from loggedin user.",
            "type": "String",
            "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Purchase request  show to sender."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/respondToPurchaseRequest": {
        "post": {
          "summary": "Respond purchase request",
          "description": "Respond to purchase request that is submitted from buyer.",
          "parameters": [
            {
            "name": "id",
            "description": "purchase request id.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "status",
            "enu": ["Accepted", "Rejected"],
            "description": "Status is accept or reject that depend on seller.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "description",
            "description": "Description about status send to buyer.",
            "type": "String",
            "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Respond to Purchase Request to buyer."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/buyTickets": {
        "post": {
          "summary": "Buy tickets",
          "description": "User can buy tickets from flipeet.",
          "parameters": [
            
            
            
            {
            "name": "tickets",
            "description": "User giver numbers of tickets.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "token",
            "description": "Charge token give from client side.",
            "type": "String",
            "in": "query"
            },
            {
            "name": "ticketAmount",
            "description": "Amount of ticket that is set from client side.",
            "type": "String",
            "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Purchase tickets successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/getTransactions": {
        "post": {
          "summary": "Get All of User",
          "description": "All buying tickets transaction or books show",
          "parameters": [
            {
            "name": "page",
            "description": "page number is send from client.",
            "type": "String",
            "in": "query"
            }
           
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "All trasaction of loggedin user."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/referToFriend": {
        "post": {
          "summary": "Send Email Invitation",
          "description": "This endpoint is used to send email ivitation to friend.",
          "parameters": [{
              "name": "email",
              "description": "Email of friend that is send inviataion.",
              "type": "String",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Send Invitaion Email Successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/editSiteInfo": {
        "post": {
          "summary": "Edit Site Content",
          "description": "This endpoint is used Edit site content like slider, email, tickets rate etc.",
          "parameters": [
            {
              "name": "id",
              "description": "by derfault id equal to 1.",
              "type": "String",
              "in": "query"
            },
            {
              "name": "breadcrumbImage",
              "description": "this is used to show image on site pages",
              "type": "String",
              "in": "query"
            },
            {
              "name": "promotionalImage",
              "description": "That images is used to promot seller",
              "type": "String",
              "in": "query"
            },
            {
              "name": "sliderImage",
              "description": "This images are slider image of home page.",
              "type": "String",
              "in": "query"
            },
            {
              "name": "contactUsEmail",
              "description": "That is used to send email from users.",
              "type": "String",
              "in": "query"
            },
            {
              "name": "ticketPrice",
              "description": "Price rate of tickets.",
              "type": "Number",
              "in": "query"
            },
            {
              "name": "promoCode",
              "description": "Promo code.",
              "type": "Number",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Site Cotent Upodated."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/getUsers": {
        "post": {
          "summary": "All user that are registered",
          "description": "All users that are registered in site is show for admins",
          "parameters": [{
              "name": "page",
              "description": "that is used for pagination",
              "type": "String",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Send Invitaion Email Successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/getPurchaseRequestsAdmin": {
        "post": {
          "summary": "All Purchase requeste",
          "description": "that is used to show all purchases request to admin.",
          "parameters": [{
              "name": "page",
              "description": "that is used for pagination",
              "type": "String",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "All Purchase request is show to admin"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/deletePurchaseRequest": {
        "post": {
          "summary": "All Purchase requeste",
          "description": "that is used to show all purchases request to admin.",
          "parameters": [{
              "name": "id",
              "description": "Purchase request id",
              "type": "String",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Delete Purchase request from admin side"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/addSliderImages": {
        "post": {
          "summary": "update Image Slider",
          "description": "This endpoint is used to update images from admin on Home screen Slider.",
          "parameters": [{
              "name": "sliderImage",
              "consumes": "multipart/form-data",
              "description": "Purchase request id",
              "type": "String",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Slider Images updated"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/editBreadcrumbImages": {
        "post": {
          "summary": "update breadcrumb Image",
          "description": "This endpoint is used to breadcrumb Image from admin on Home screen Slider.",
          "parameters": [{
              "name": "breadcrumbImage",
              "consumes": "multipart/form-data",
              "description": "Purchase request id",
              "type": "String",
              "in": "query"
            }

          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "breadcrumb Images updated"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/getSiteContent": {
        "post": {
          "summary": "Get Site Content",
          "description": "This endpoint is used get Site content like images, sliderImages, tickets amount etc.",
          "parameters": [
            {
              "name": "breadcrumbImage",
              "consumes": "multipart/form-data",
              "description": "Purchase request id",
              "type": "String",
              "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Get Site Content"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/activities/blog_post/add": {
        "post": {
          "summary": "Create Blog",
          "description": "Blog is created by only admin to show users.",
          "parameters": [
            {
              "name": "title",
              "description": "Title of blog",
              "type": "String",
              "in": "query"
            }, {
              "name": "description",
              "description": "Description of blog",
              "type": "String",
              "in": "query"
            },
            {
              "name": "image",
              "consumes": "multipart/form-data",
              "description": "Blog of Image",
              "type": "String",
              "in": "query"
            }
          ],
          "tags": [
            "Activities"
          ],
          "responses": {
            "200": {
              "description": "Blog is Created."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
    }
  }